Fix version file generation in external git repo
If the source tarball is extracted within a git repository, the
version file generation code would try to re-create the version
files and would very likely fail doing so.

This is an issue on Arch Linux, where packages are apparently
built from within a repo that hold the package build information.