Produce deterministic inode numbers (needed to fix #91)
While most of the code typically ensures that elements are kept
in a deterministic order, the code that assigned inode numbers
was iterating a hash table, which by itself guaranteed FIFO
semantics, but items were inserted from multiple threads when
scanning the input file system.

This change adds a sorting step before assigning inode numbers.
(This shouldn't be much of a performance hit.)