Fix heap-use-after-free when writing section index
When writing the section index block, an additional entry was added
to the index, potentially reallocating the vector containing the
index. However, we previously took the address of the vector data
in order to write the index, so that address is now invalid.

Fix is by not adding the extra entry to the index.