fix(file_scanner): unexpected "inode has no file" (fixes gh #217)
When introducing an optimization to skip hashing of large files if they
already differ in the first 4 KiB, only `unique_size_` was updated to
be keyed on (size, start_hash). However, the same change is necessary
for `first_file_hashed_`, as there can otherwise be collisions if files
with the same size, but different start hashes are processed at the same
time.