fix(lzma): lzma options pointing at wrong instance (fixes gh #224)
When cloning an `lzma_block_compressor`, the cloned `filters_` would
point to the original `opt_lzma_` instead of the cloned one. While
this is obviously a bug, the code would usually work and only
occasionally bail out with an lzma error.

This change does away with having `filters_` as a member and rather
builds the array each time `compress` is called. The overhead is
completely negligible.