fix(utils_test): u8string triggers linker error with non-C++20 gtest
This is a long standing issue and it's not going to go away, as package
manager prefer that stuff isn't fetched at build time. So we just check
if certain gtest features work with the system-installed version and
work around missing features downstream. Out of the box, with the custom
built ctest, we'll still get all the good stuff. If the system-installed
version is preferred, we probably have some degradation in test output.