feat(filesystem_parser): use optimized search for section header
Turns out that `std::boyer_moore_searcher` is pretty slow for our small
needle. This custom searcher assumes that the major version byte is the
most unlikely to occur (at least for now) and uses `strchr` to search
for that (in the hope that this is nicely optimized), and then validate
any match against the full magic.

Here's a benchmark (x86_64) for a tiny image with a 110 MiB header:

```
$ hyperfine "./dwarfsck.orig --no-check -q td.dwarfs" "./dwarfsck --no-check -q td.dwarfs"
Benchmark 1: ./dwarfsck.orig --no-check -q td.dwarfs
  Time (mean ± σ):      72.5 ms ±   4.0 ms    [User: 64.5 ms, System: 7.8 ms]
  Range (min … max):    67.1 ms …  84.2 ms    35 runs

Benchmark 2: ./dwarfsck --no-check -q td.dwarfs
  Time (mean ± σ):      16.0 ms ±   0.9 ms    [User: 8.6 ms, System: 7.3 ms]
  Range (min … max):    14.4 ms …  19.5 ms    161 runs

Summary
  ./dwarfsck --no-check -q td.dwarfs ran
    4.54 ± 0.35 times faster than ./dwarfsck.orig --no-check -q td.dwarfs
```