Move dwarfs manpage to section 8 (sysadmin)
dwarfs lives in /usr/sbin and is considered a sysadmin program.
Its manpage should live in section 8.

This would also be consistent with similar tools like fuse(8)
or mount(8).

debian's lintian recognises this discrepancy as it inspects
the 'dwarfs' package and produces a mild pedantic notice.

Tested with `ninja install` output, `dwarfs --man` and `cpack`
archive.