build: only create symbolic mount links at install time
in the build directory were no longer pointing at a target, which was a
bit irritating. As a result, `ninja` (maybe also `make`) didn't recognize
this as a target that had been successfully created and re-created the
links over and over again.

There's really no need for these links to be present in the build dir,
so this change just moves link creation to install time.