fix(metadata_types): off-by-one error when recovering `self_index` fields
This fixes the off-by-one bug, which can cause the sentinel directory
to have a non-zero `self_entry`. This index is never actually used, so
the bug is relatively harmless. The non-zero `self_entry` does however
cause the metadata consistency checks to fail.

So beyond the actual fix, this change adds somewhat stricter checking
(making sure we recover *all* `self_entry` fields), but at the same
time relaxes the consistency check to only emit a warning when a non-zero
self entry is found in the sentinel.

Additionally, when the `metadata_builder` sees a non-zero self entry in
the sentinel, it also logs a brief message and fixes the entry.