feat(filesystem_v2): add `entries_in_data_order`
This allows to split the preparation phase (ordering all entries) and the
actual walk. It also conveniently allows to use a regular range-based
loop instead of a callback. Last but not least, you can re-use the same
iterable to walk in data order multiple times without repeating the
prepartion step.